requirements.txt
python-doctr[tf]
tf2onnx
flask
certifi
urllib3==1.24.3